requirements.txt
<<<<<<< HEAD
asgiref==3.8.1
Django==5.1.4
pillow==11.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
djangorestframework==3.14.0
=======
asgiref==3.8.1
Django==5.1.4
pillow==11.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
>>>>>>> e86b7b9f37d611aad436361104d1f4cb0dea76e7
